+++ .github/workflows/push.yml
@@ -25,9 +25,9 @@ jobs:
         run: poetry run mypy . --exclude tests
       - name: run 3.9- tests
         if: ${{ matrix.python-version }} != 3.10
-        run: poetry run pytest --ignore tests/test_together.py
+        run: poetry run pytest --ignore tests/test_310.py
       - name: run 3.10+ tests
         if: ${{ matrix.python-version }} == 3.10
-        run: poetry run pytest --ignore tests/test_together.py
+        run: poetry run pytest
       - name: linting
-        run: poetry run flake8 --exclude tests/test_together.py
+        run: poetry run flake8 --exclude tests/test_310.py
